Rename locust.events module to locust.event Introduce global instance of Events() stored in locust.events which can be used to register event listeners at the module level of locustfiles. Remove some unused imports.
