+++ docs/extending-locust.rst
@@ -4,30 +4,18 @@ Extending Locust
 
 Locust comes with a number of events hooks that can be used to extend Locust in different ways.
 
-Most event hooks live on the Environment instance under the :py:attr:`events <locust.env.Environment.events>` 
-attribute. However the Environment instance is not available at the module level of your test 
-file, so in order to utilize the :py:attr:`events <locust.env.Environment.events>` on the 
-:py:class:`Environment <locust.env.Environment>` class, you can utilize the :py:attr:`locust.events.init` 
-event which is triggered when Locust is started.
+Event hooks live on the Environment instance under the :py:attr:`events <locust.env.Environment.events>` 
+attribute. However, since the Environment instance hasn't been created when locustfiles are imported,  
+the events object can also be accessed at the module level of the locustfile through the 
+:py:attr:`locust.events` variable.
 
-Adding the following code at the module level of you locustfile.py will print a message when Locust starts::
+Here's an example on how to set up an event listener::
 
-    from locust.events import init
+    from locust import events
     
-    @init.add_listener
-    def on_locust_init(environment, **kw):
-        print("Locust is starting")
-
-
-We can use this event listener to setup listeners for other events::
-
-    from locust.events import init
-    
-    @init.add_listener
-    def on_locust_init(environment, **kw):
-        @environment.events.request_success
-        def my_success_handler(request_type, name, response_time, response_length, **kw):
-            print("Successfully made a request to: %s" % name)
+    @events.request_success.add_listener
+    def my_success_handler(request_type, name, response_time, response_length, **kw):
+        print("Successfully made a request to: %s" % name)
 
 
 .. note::
@@ -46,13 +34,14 @@ Adding Web Routes
 ==================
 
 Locust uses Flask to serve the web UI and therefore it is easy to add web end-points to the web UI.
-From the Environment instance, we can access the Flask app and set up a new route::
+By listening to the :py:attr:`init <locust.event.Events.init>` event, we can retrieve a reference 
+to the Flask app instance and use that to set up a new route::
 
-    from locust.events import init
+    from locust import events
     
-    @init.add_listener
-    def on_locust_init(environment, **kw):
-        @environment.web_ui.app.route("/added_page")
+    @events.init.add_listener
+    def on_locust_init(web_ui, **kw):
+        @web_ui.app.route("/added_page")
         def my_added_page():
             return "Another page"
