+++ locust/clients.py
@@ -9,7 +9,6 @@ from requests.exceptions import (InvalidSchema, InvalidURL, MissingSchema,
 
 from urllib.parse import urlparse, urlunparse
 
-from . import events
 from .exception import CatchResponseError, ResponseError
 
 absolute_http_url_regexp = re.compile(r"^https?://", re.I)
