+++ locust/contrib/fasthttp.py
@@ -15,7 +15,6 @@ from gevent.timeout import Timeout
 from geventhttpclient.useragent import UserAgent, CompatRequest, CompatResponse, ConnectionError
 from geventhttpclient.response import HTTPConnectionClosed
 
-from locust import events
 from locust.core import Locust
 from locust.exception import LocustError, CatchResponseError, ResponseError
