+++ locust/core.py
@@ -14,7 +14,6 @@ from gevent import GreenletExit, monkey
 # See: https://github.com/requests/requests/issues/3752#issuecomment-294608002
 monkey.patch_all()
 
-from . import events
 from .clients import HttpSession
 from .exception import (InterruptTaskSet, LocustError, RescheduleTask,
                         RescheduleTaskImmediately, StopLocust, MissingWaitTimeError)
@@ -169,7 +168,7 @@ class Locust(object):
                     logger.error("%s\n%s", e, traceback.format_exc())
             if hasattr(self, "teardown") and self._teardown_is_set is False:
                 self._set_teardown_flag()
-                events.quitting.add_listener(self.teardown)
+                self.environment.events.quitting.add_listener(self.teardown)
 
     @classmethod
     def _set_setup_flag(cls):
@@ -373,7 +372,7 @@ class TaskSet(object, metaclass=TaskSetMeta):
                     logger.error("%s\n%s", e, traceback.format_exc())
             if hasattr(self, "teardown") and self._teardown_is_set is False:
                 self._set_teardown_flag()
-                events.quitting.add_listener(self.teardown)
+                self.environment.events.quitting.add_listener(self.teardown)
 
     @classmethod
     def _set_setup_flag(cls):
